+++ src/editor-model/commands-delete.ts
@@ -2,48 +2,47 @@ import { register } from '../editor/commands';
 import type { ModelPrivate } from './model-private';
 import { deleteForward, deleteBackward, deleteRange } from './delete';
 import { wordBoundaryOffset } from './commands';
-import { contentWillChange } from './listeners';
 
 register(
   {
     deleteAll: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteContent' }) &&
+      model.contentWillChange({ inputType: 'deleteContent' }) &&
       deleteRange(model, [0, -1], 'deleteContent'),
     deleteForward: (model: ModelPrivate): boolean => deleteForward(model),
     deleteBackward: (model: ModelPrivate): boolean => deleteBackward(model),
     deleteNextWord: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteWordForward' }) &&
+      model.contentWillChange({ inputType: 'deleteWordForward' }) &&
       deleteRange(
         model,
         [model.anchor, wordBoundaryOffset(model, model.position, 'forward')],
         'deleteWordForward'
       ),
     deletePreviousWord: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteWordBackward' }) &&
+      model.contentWillChange({ inputType: 'deleteWordBackward' }) &&
       deleteRange(
         model,
         [model.anchor, wordBoundaryOffset(model, model.position, 'backward')],
         'deleteWordBackward'
       ),
     deleteToGroupStart: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteSoftLineBackward' }) &&
+      model.contentWillChange({ inputType: 'deleteSoftLineBackward' }) &&
       deleteRange(
         model,
         [model.anchor, model.offsetOf(model.at(model.position).firstSibling)],
         'deleteSoftLineBackward'
       ),
     deleteToGroupEnd: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteSoftLineForward' }) &&
+      model.contentWillChange({ inputType: 'deleteSoftLineForward' }) &&
       deleteRange(
         model,
         [model.anchor, model.offsetOf(model.at(model.position).lastSibling)],
         'deleteSoftLineForward'
       ),
     deleteToMathFieldStart: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteHardLineBackward' }) &&
+      model.contentWillChange({ inputType: 'deleteHardLineBackward' }) &&
       deleteRange(model, [model.anchor, 0], 'deleteHardLineBackward'),
     deleteToMathFieldEnd: (model: ModelPrivate): boolean =>
-      contentWillChange(model, { inputType: 'deleteHardLineForward' }) &&
+      model.contentWillChange({ inputType: 'deleteHardLineForward' }) &&
       deleteRange(model, [model.anchor, -1], 'deleteHardLineForward'),
   },
   {